Add a new failover option prioritising nodes
This makes it possible to define a list of nodes you'd prefer to communicate
with in cluster mode.

This behaves similarly to FAILOVER_DISTRIBUTE, where it randomises the list of
nodes before trying them in sequence - however it takes that random list and
sorts any preferred nodes to the top first. All candidate nodes are still in
the list, but the preferred ones get tried first.

This is extremely helpful for setups where you know which nodes are closer
(i.e. not crossing an AZ boundary).